requirements.txt
# Dependencies for connecting with the broker
MetaTrader5==5.0.45   # Library to connect with MetaTrader 5 trading platform
pydantic==1.10.7      # Data validation and settings management using Python type annotations

# Dependencies for working with data
pandas==2.0.1         # Powerful data manipulation and analysis library

# Dependencies for feature extraction
ta==0.10.2            # Technical analysis library for feature extraction

# Dependencies for machine learning and neural networks
scikit-learn==1.3.0   # Machine learning library for data mining and analysis
tensorflow==2.13.0   # Deep learning framework by Google
keras==2.13.1        # High-level neural networks API running on top of TensorFlow

# Dependencies for reinforcement learning
gym==0.26.2          # OpenAI Gym, a toolkit for developing and comparing reinforcement learning algorithms
stable-baselines3[extra]==2.0.0   # Reinforcement learning algorithms implementation based on OpenAI Baselines, with extra dependencies

# Dependency for working with Parquet files
pyarrow==12.0.1      # Library for working with columnar data stored in Parquet format

PyQt6==6.5.3 # To Draw Figures on an external GUI

mplfinance==0.12.10b0 # To plot candlestick charts

pytest==7.4.2 # Check if code work

python-dotenv==1.0.0 # Read env file